updated as per comments, used ginko instead of testify, shorten the duration of tests
